+++ SearchACT.py
@@ -22,267 +22,33 @@ Version log:
     - multiple rule syntax
     - calculator re-implement
     - search key by string matching
+- 1.8.0:
+    - new structure
 
 Next version
 - include removing function
 
 """
 
-__version__ = '1.7.0'
+__version__ = '1.8.0'
 
 
 from sys import argv
 from os import path
 import time
 ### coustom functions
-from modules.tools import save_obj_to_pickle, read_pickle
+from modules.search import SearchACT
+from modules.calculator import parse
+from modules.contact import Contact
+from modules.help_info import help_
 ###
 
 PATH_OF_SCRIPT = path.dirname(argv[0])
 PATH_DICT_KEY_CONTACT = path.join(PATH_OF_SCRIPT, '_dict_key_contacts.pickle')
 PATH_DICT_TERM_KEY = path.join(PATH_OF_SCRIPT, '_dict_terms_key.pickle')
-DICT_KEY_CONTACT = read_pickle(PATH_DICT_KEY_CONTACT)
-DICT_TERM_KEY = read_pickle(PATH_DICT_TERM_KEY)
+
 LI_EXIST_WORDS = ['exit', 'bye', 'ex', 'bye bye', 'quit']
 
-def add_searchTerm_to_key():
-    #global DICT_KEY_CONTACT, DICT_TERM_KEY
-
-    str_input_KEY = input('Type the E-mail ID (for example, type "benlin" for "benlin@actgenomics.com") or type "exit" to exit. >>> ').lower()
-    if str_input_KEY in LI_EXIST_WORDS:
-        return 0
-    elif str_input_KEY not in DICT_KEY_CONTACT.keys():
-        print('*** The ID not exist. ***')
-        return 0
-    str_input_Term = input('Type the new search term or type "exit" to exit.\n>>> ').lower()
-    if str_input_Term in LI_EXIST_WORDS:
-        return 0
-
-    DICT_TERM_KEY.setdefault(str_input_Term.lower(), set()).add(str_input_KEY)
-    save_obj_to_pickle(path.splitext(PATH_DICT_TERM_KEY)[0], DICT_TERM_KEY)
-    print(f'Successfuly add new search terms. "{str_input_Term}":"{str_input_KEY}"')
-    return 0
-
-def add_contactInfo():
-    str_input_KEY = input('Type the E-mail ID (for example, type "benlin" for "benlin@actgenomics.com") or type "exit" to exit.\n>>> ').lower()
-    if str_input_KEY in LI_EXIST_WORDS:
-        return 0
-    elif str_input_KEY not in DICT_KEY_CONTACT.keys():
-        print('*** The ID not exist. ***')
-        return 0
-
-    str_add_info = input('Type the new info of the person or type "exit" to exit.\n>>> ').lower()
-    if str_add_info in LI_EXIST_WORDS:
-        return 0
-    DICT_KEY_CONTACT[str_input_KEY].append(str_add_info)
-    save_obj_to_pickle(path.splitext(PATH_DICT_KEY_CONTACT)[0], DICT_KEY_CONTACT)
-    print(f'Successfuly add new info terms. "{DICT_KEY_CONTACT[str_input_KEY]}"')
-    return 0
-    
-def search(str_input):
-    '''
-    Return by the matched keys by str_input.
-
-    '''
-    set_match_key = set()
-    for term in DICT_TERM_KEY:
-        if str_input in term:
-            for key in DICT_TERM_KEY[term]:
-                set_match_key.add(key)
-    return set_match_key
-
-def index_from_end(li_, match):
-    for i in range(len(li_)-1, 0, -1):
-        if li_[i] == match:
-            return i
-    return None
-
-def cal_multi_condi_li(li_ele):
-    #print(li_ele)
-    li_ele = [x for x in li_ele if x]
-    if len(li_ele)==1 and type(li_ele[0])==str:
-        return search(li_ele[0])
-    elif len(li_ele)==1 and type(li_ele[0])==set:
-        return li_ele[0]
-    elif len(li_ele)>1 and not [x for x in li_ele if x in ['|', '&']]:
-        print('should not exist')
-        set_ = set()
-        for i in li_ele:
-            set_.update(search(i))
-        return set_
-    else:
-        for i in [x for x in li_ele if x in ['|', '&']]:
-            #offset = li_ele.index(i)
-            offset = index_from_end(li_ele, i)
-            left, op, right = li_ele[:offset], li_ele[offset:offset+1], li_ele[offset+1:]
-            if op == ['&']:
-                return cal_multi_condi_li(left) & cal_multi_condi_li(right)
-            elif op == ['|']:
-                return cal_multi_condi_li(left) | cal_multi_condi_li(right)
-
-def parse_formula(s):
-    if s == '':
-        return set()
-    l = []
-    r = []
-    s = s.replace(' ','')
-    li_element = []
-    temp = ''
-    for e in s:
-        if e == '(':
-            li_element.append(temp)
-            temp = ''
-            l.append(len(li_element))
-        elif e == ')':
-            li_element.append(temp)
-            temp = ''
-            r.append(len(li_element))
-            li_element[l[-1]:r[-1]+1] = [cal_multi_condi_li(li_element[l[-1]:r[-1]+1])] + ['']*(r[-1]+1 - l[-1] - 1)
-            l.pop()
-            r.pop()
-        elif e in '&|':
-            li_element.append(temp)
-            temp = ''
-            li_element.append(e)
-        else:
-            temp += e
-    li_element.append(temp)    
-    return cal_multi_condi_li(li_element)
-
-""" # ben's modified --> fix some issue 
-
-def cal(s):
-    s = s.replace(' ', '')
-    if not any(x in s for x in ['+','-','*', '/']):
-        return float(s)
-    for i in [x for x in s if x in ['+','-','*', '/']]:
-        left, op, right = s.partition(i)
-        if op in ['*', '/','+','-']: 
-            if op == '*':
-                return(cal(left) * cal(right))
-            elif op == '/':
-                return(cal(left) / cal(right))
-            elif op == '+':
-                return(cal(left) + cal(right))
-            elif op == '-':
-                return(cal(left) - cal(right))
-def parse(s):
-    if s == '':
-        return 'Empty string!'
-    l = []
-    r = []
-    s = s.replace(' ','')
-    
-    for p in range(len(s)):
-        if any(x in s for x in ['(', ')']):   
-            if s[p] == '(':
-                l.append(p)
-            elif s[p] == ')':
-                if len(l) == 0:
-                    print('left bracket first')
-                else:
-                    r.append(p)
-                    ans = cal(s[l[-1]+1:r[-1]])
-                    len_bra = r[-1] - l[-1] +1
-                    s = "".join((s[:l[-1]], str(ans)+' '*(len_bra-len(str(ans))), s[r[-1]+1:]))
-                    l.pop()
-                    r.pop()
-        
-    return str(cal(s))
-"""
-def cal(s,lop):
-    s = handle(s)
-    if not any(x in s for x in ['+','-','*', '/']):
-        return float(s)
-    for i in ['+','-','*', '/']:
-        left, op, right = s.partition(i)
-        # print(i, [left, op, right])
-        if op in ['+', '-','*','/']:
-            if op == '*':
-                if right == '':
-                    return(cal(left,lop))
-                else:
-                    return(cal(left,lop) * cal(right,lop))
-            elif op == '/':
-                if lop != '/':
-                    lop = '/'
-                elif lop == '/':
-                    return(cal(left,lop) * cal(right,lop))   
-                    lop = ''
-                return(cal(left,lop) / cal(right,lop))
-            elif op == '+':
-                if left == '':
-                    return(cal('0+'+right,lop))
-                else:
-                    return(cal(left,lop) + cal(right,lop))
-            elif op == '-':
-                if left == '':
-                    return(cal('0-'+right,lop))
-                else:
-                    if lop != '-':
-                        lop = '-'
-                    elif lop == '-':
-                        return(cal(left,lop) + cal(right,lop))   
-                        lop = ''
-                    return(cal(left,lop) - cal(right,lop))
-
-
-def parse(s,lop=''):
-    if s == '':
-        return 'Empty string!'
-    l = []
-    r = []
-    s = s.replace(' ','')
-    while any(x in s for x in ['(',')']):
-        for p in range(len(s)):
-            if s[p] == '(' and len(r) < 1:
-                l.append(p)
-            elif s[p] == ')':
-                if len(l) == 0:
-                    return('left bracket first!')
-                else:
-                    r.append(p)
-                    if s[l[-1]+1:r[-1]] == '':
-                        return 'formula error!'
-                    else:
-                        ans = cal(s[l[-1]+1:r[-1]],lop)
-                    s = "".join((s[:l[-1]],str(ans),s[r[-1]+1:]))
-                    l.clear()
-                    r.pop()
-                    break
-    s = handle(s)
-    return str(cal(s,lop))
-
-def handle(s):  
-    timestart = time.time()
-    while any(x in s for x in ['+-','--','*-', '/-']) and time.time() < timestart+5:
-        if '+-' in s:
-            s=s.replace('+-','-')
-        elif '--' in s:
-            if s.find('--') == 0:
-                s=s.replace('--','')
-            s=s.replace('--','+')
-        elif '*-' in s:
-            p = s.find('*-')
-            for i in range(p-1,-1,-1):
-                if s[i] in ['+','-','*','/']: 
-                    s="".join((s[:i+1],'-',s[i+1:]))
-                    s="".join((s[:p+1],'*',s[p+3:]))
-                    break
-                elif i == 0:
-                    s="".join(('-',s[i:]))
-                    s="".join((s[:p+1],'*',s[p+3:]))
-        elif '/-' in s:
-            p = s.find('/-')
-            for i in range(p-1,-1,-1):
-                if s[i] in ['+','-','*','/']: 
-                    s="".join((s[:i+1],'-',s[i+1:]))
-                    s="".join((s[:p+1],'/',s[p+3:]))
-                    break
-                elif i == 0:
-                    s="".join(('-',s[i:]))
-                    s="".join((s[:p+1],'/',s[p+3:]))
-    return(s)
 
 # main process
 def main():
@@ -290,6 +56,8 @@ def main():
 #    global DICT_KEY_CONTACT, DICT_TERM_KEY,  PATH_DICT_KEY_CONTACT, PATH_DICT_TERM_KEY    
     while True:
         str_input = input('\nSeach >>> ')#.lower()
+        Contact_ = Contact(PATH_DICT_KEY_CONTACT, PATH_DICT_TERM_KEY)
+        SearchACT_ = SearchACT(Contact_.DICT_TERM_KEY) 
         '''
         if str_input in DICT_TERM_KEY:
             ## matched
@@ -299,7 +67,16 @@ def main():
         # add search term function
         if str_input in ['*addterm', '*add1']:
             print('into adding mode')
-            ans = add_searchTerm_to_key()
+            str_input_KEY = input('Type the E-mail ID (for example, type "benlin" for "benlin@actgenomics.com") or type "exit" to exit. >>> ').lower()
+            if str_input_KEY in LI_EXIST_WORDS:
+                break
+            elif str_input_KEY not in Contact_.DICT_KEY_CONTACT.keys():
+                print('*** The ID not exist. ***')
+                continue
+            str_input_Term = input('Type the new search term or type "exit" to exit.\n>>> ').lower()
+            if str_input_Term in LI_EXIST_WORDS:
+                break 
+            ans = Contact_.add_searchTerm_to_key(str_input_Term, str_input_KEY)
             if ans == 0:
                 pass
             else:
@@ -307,8 +84,17 @@ def main():
 
         # add contact info function
         elif str_input in ['*addinfo', '*add2']:
-            print('into editing mode')
-            ans = add_contactInfo()
+            str_input_KEY = input('Type the E-mail ID (for example, type "benlin" for "benlin@actgenomics.com") or type "exit" to exit.\n>>> ').lower()
+            if str_input_KEY in LI_EXIST_WORDS:
+                break
+            elif str_input_KEY not in Contact_.DICT_KEY_CONTACT.keys():
+                print('*** The ID not exist. ***')
+                continue
+            str_add_info = input('Type the new info of the person or type "exit" to exit.\n>>> ').lower()
+            if str_add_info in LI_EXIST_WORDS:
+                break
+
+            ans = Contact_.add_contactInfo(str_input_KEY, str_add_info)
             if ans == 0:
                 pass
             else:
@@ -316,28 +102,10 @@ def main():
         # remove contact info function
         elif str_input in ['*rminfo', '*rm2']:
             pass
-            """
-            print('into editing mode')
-            ans = add_contactInfo()
-            if ans == 0:
-                pass
-            else:
-                break
-            """
-
+            #Contact_.rm_ifno()
         elif str_input in ['*rebuild', '*rb']:
             pass
-            '''
-            import updata_contact
-            print('** this operation will wipe out the term created by your own. Sure?')
-            str_input = input('\n(Y/N) >>> ').lower()
-            if str_input == 'y':
-                updata_contact.MakeDict(updata_contact.locate_contact_file())
-                DICT_KEY_CONTACT = read_pickle(PATH_DICT_KEY_CONTACT)
-                DICT_TERM_KEY = read_pickle(PATH_DICT_TERM_KEY)
-            else:
-                pass
-            '''
+            #Contact_.re_build()
         elif str_input.startswith('$'):
             print(eval(str_input[1:]))
         elif str_input in LI_EXIST_WORDS:
@@ -358,20 +126,23 @@ def main():
                         print('formula error !')
                         pass
         elif str_input in ['*help', '*h']:
-            print('\n# Type "*addterm", "*add1" to add search term linked to someone. \n# Type "*addinfo" or "*add2" to add additional info to someone. \n# Type "exit", "ex" to leave.')
+            print(f' version:{__version__}')
+            print(help_())
         else:
             #raw_match(str_input, DICT_TERM_KEY, DICT_KEY_CONTACT)
             try:
                 #set_matches = search(str_input)
-                set_matches = parse_formula(str_input)
+                #set_matches = parse_formula(str_input)
+                set_matches = SearchACT_.parse_formula(str_input)
             except:
                 print('formula error !')
+                set_matches = None
                 pass
             if not set_matches:
                 print(f'\n "{str_input}" not found. Retry or type "exit" to exit.')
             else:
                 for r in set_matches:
-                    print('\n', '>'+'\t'.join(DICT_KEY_CONTACT[r]))
+                    print('\n', '>'+'\t'.join(Contact_.DICT_KEY_CONTACT[r]))
