--- modules/tools.py
@@ -1,10 +1,0 @@
-import pickle
-
-def save_obj_to_pickle(path, obj): # will end with pickle
-    print('# saving file to:', '%s.pickle'%path)
-    with open('%s.pickle'%path, 'wb') as f_pkl:
-        pickle.dump(obj, f_pkl)
-
-def read_pickle(path_):
-    with open(path_, 'rb') as file:
-        return pickle.load(file)
